# docs/swagger.yaml
basePath: /
definitions:
...
  httpx.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  httpx.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/httpx.ErrorResponse'
      success:
        type: boolean
    type: object
  mfa.TOTPVerifyRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
info:
  contact: {}
  description: Auth API for devclub services.
  title: Kael API
  version: "1.0"
paths:
...
  /mfa/totp/enable:
    post:
      description: Generate a new TOTP secret and return the secret and otpauth URL
        for QR code
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Response'
            - properties:
                error:
                  $ref: '#/definitions/httpx.ErrorResponse'
              type: object
      summary: Enable TOTP
      tags:
      - mfa
  /mfa/totp/verify:
    post:
      consumes:
      - application/json
      description: Verify the first TOTP code to enable TOTP MFA for the user
      parameters:
      - description: TOTP verification payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mfa.TOTPVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Response'
            - properties:
                data:
                  additionalProperties:
                    type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Response'
            - properties:
                error:
                  $ref: '#/definitions/httpx.ErrorResponse'
              type: object
      summary: Verify TOTP
      tags:
      - mfa
  /oauth/github/link:
    get:
      description: Link GitHub account to the currently authenticated user
      parameters:
      - description: Device Identifier
        in: header
        name: X-Device-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: Link GitHub
      tags:
      - oauth
  /oauth/github/start:
    get:
      description: Get the GitHub OAuth authorization URL
      parameters:
      - description: Device Identifier
        in: header
        name: X-Device-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: Start GitHub OAuth
      tags:
      - oauth
  /oauth/google/link:
    get:
      description: Link Google account to the currently authenticated user
      parameters:
      - description: Device Identifier
        in: header
        name: X-Device-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: Link Google
      tags:
      - oauth
  /oauth/google/start:
    get:
      description: Get the Google OAuth authorization URL
      parameters:
      - description: Device Identifier
        in: header
        name: X-Device-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: Start Google OAuth
      tags:
      - oauth
...
swagger: "2.0"